Update config and layer media types when OCIfying
Prior to this change, quay.io would reject OCIfied images because they
contained Docker-type base layers and configs.

With this change, instead of simply appending to the base image, we
collect base layers (OCIfying them if necessary) then append them *and*
the new layers, on top of an empty base image.

This also mutates the image's config media type (for the same reason),
and carries forward the base image's annotations, including annotations
about the base image that we collect earlier in the build process.