Allow specification of the binary name/path
Some container usage expects to run an executable in a specific
location, and some projects aren't well laid out to generate binaries
with the expected names.

This allows the binary name to be specified in configuration (in a
similar way to goreleaser), which should allow for easier migration from
other build mechanisms (and usage patterns of their images) to `ko`.

This does not affect the name or path for Windows builds - the path
structure is fundamentally different, and I think that if it's desirable
to change the name/path for them, we should use a separate config
argument.