Update awslabs/amazon-ecr-credential-helper to latest
This update fixes an apparent conflict between the
amazon-ecr-credential-helper library and the aws-sdk-go-v2 library,
introduced when aws-sdk-go-v2 was updated as a byproduct of a cosign
upgrade in #1191.

This mismatch resulted in an error:
"ecr: Failed to get authorization token: not found, ResolveEndpointV2"

However, this error was not readily apparent due to overridding the
ecr-login module's logger to io.Discard in pkg/commands/config.go.

Resolves: #1245
Signed-off-by: Eric Searcy <eric@linuxfoundation.org>