Update recorder to lazy create file.
Currently the resolver command creates the image-refs file on
initilization. This causes git to be dirty during builds. This change moves the
file creation to be in the recorder itself and on the first time the Publish()
method is called. This happens after the build so git is clean. This will mean
that any errors on file creation will be reported after the build rather than
before.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>