Change OCI Layout publisher to lazy create layout after build.
When using ko with OCI Layout, Git status will always be dirty because the
layout is created before this build. This changes the layout to be created
after the build so that git will be clean. The means that any error creating
the layout won't be seen until after the build.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>